TEIID-4545: add phoenix/hbase support functions, joins
