+++ connectors/jdbc/translator-hbase/src/main/resources/META-INF/services/org.teiid.translator.ExecutionFactory
@@ -1,1 +1,1 @@
-org.teiid.translator.hbase.HBaseExecutionFactory
+org.teiid.translator.hbase.PhoenixExecutionFactory
