+++ teiid-feature-pack/wildfly-integration-feature-pack/src/main/resources/content/bin/scripts/teiid-domain-mode-install.cli
@@ -128,7 +128,7 @@ connect
 /profile=ha/subsystem=teiid/translator=excel:add(module=org.jboss.teiid.translator.excel)
 /profile=ha/subsystem=teiid/translator=impala:add(module=org.jboss.teiid.translator.hive)
 /profile=ha/subsystem=teiid/translator=prestodb:add(module=org.jboss.teiid.translator.prestodb)
-/profile=ha/subsystem=teiid/translator=hbase:add(module=org.jboss.teiid.translator.hbase)
+/profile=ha/subsystem=teiid/translator=phoenix:add(module=org.jboss.teiid.translator.hbase)
 /profile=ha/subsystem=teiid/translator=hana:add(module=org.jboss.teiid.translator.jdbc)
 /profile=ha/subsystem=teiid/translator=vertica:add(module=org.jboss.teiid.translator.jdbc)
 /profile=ha/subsystem=teiid/translator=actian-vector:add(module=org.jboss.teiid.translator.jdbc)
@@ -206,7 +206,7 @@ connect
 /profile=full-ha/subsystem=teiid/translator=excel:add(module=org.jboss.teiid.translator.excel)
 /profile=full-ha/subsystem=teiid/translator=impala:add(module=org.jboss.teiid.translator.hive)
 /profile=full-ha/subsystem=teiid/translator=prestodb:add(module=org.jboss.teiid.translator.prestodb)
-/profile=full-ha/subsystem=teiid/translator=hbase:add(module=org.jboss.teiid.translator.hbase)
+/profile=full-ha/subsystem=teiid/translator=phoenix:add(module=org.jboss.teiid.translator.hbase)
 /profile=full-ha/subsystem=teiid/translator=hana:add(module=org.jboss.teiid.translator.jdbc)
 /profile=full-ha/subsystem=teiid/translator=vertica:add(module=org.jboss.teiid.translator.jdbc)
 /profile=full-ha/subsystem=teiid/translator=actian-vector:add(module=org.jboss.teiid.translator.jdbc)
