+++ teiid-feature-pack/wildfly-integration-feature-pack/src/main/resources/content/bin/scripts/teiid-standalone-ha-mode-install.cli
@@ -85,7 +85,7 @@ connect
 /subsystem=teiid/translator=excel:add(module=org.jboss.teiid.translator.excel)
 /subsystem=teiid/translator=impala:add(module=org.jboss.teiid.translator.hive)
 /subsystem=teiid/translator=prestodb:add(module=org.jboss.teiid.translator.prestodb)
-/subsystem=teiid/translator=hbase:add(module=org.jboss.teiid.translator.hbase)
+/subsystem=teiid/translator=phoenix:add(module=org.jboss.teiid.translator.hbase)
 /subsystem=teiid/translator=hana:add(module=org.jboss.teiid.translator.jdbc)
 /subsystem=teiid/translator=vertica:add(module=org.jboss.teiid.translator.jdbc)
 /subsystem=teiid/translator=actian-vector:add(module=org.jboss.teiid.translator.jdbc)
